add api error model
