+++ src/api/schemas/register_user.rs
@@ -1,6 +1,6 @@
-use crate::{
-    application::ApiError,
-    model::register_user::{RegisterUserRequest, UserEmail, UserPassword},
+use crate::model::{
+    api_error::ApiError,
+    register_user::{RegisterUserRequest, UserEmail, UserPassword},
 };
 use anyhow::Result;
 use serde::Deserialize;
